hw/riscv: iommu: Add PRGI to the page request payload
The PCI specification states that a page request payload carries a page
request group index so that the response that is later received through
the PRGR notification will be able to match the request based on the
index.

By adding the group index to the payload, we follow the PCI spec and
allow the callers to submit multiple requests in parallel.

Signed-off-by: Sebastien Boeuf <seb@rivosinc.com>