hw/riscv: iommu: Carry the PRGI through IOVA
When the PRGR reaches the IOMMU, there's one more step for the IOMMU to
let the PCI device know about it. The IOMMUTLBEvent contains an entry of
type IOMMUTLBEntry that itself carries an IOVA field. Given the IOVA is
useless for PRGR given a PRGR payload doesn't carry an IOVA, we overload
it with the group index (PRGI) from the payload.

Signed-off-by: Sebastien Boeuf <seb@rivosinc.com>