hw/riscv: iommu: Allow access flags through page request
Let the caller decide the status of the flags that should be sent with
the request. The flags are access flags read-only (RO), write-only (WO)
and read-write (RW) combined with the last request flag indicating if
this request is the last from the given Page Request Group (PRG).

Given the implementation doesn't handle requests in several steps, we
expect the caller to set the last-page bit for each request.

Signed-off-by: Sebastien Boeuf <seb@rivosinc.com>