hw/riscv/virt.c: add 'legacy-aia-dt' compat option
Recent changes in the AIA DT of the 'virt' machine turned out to be
quite disrupting for any available firmware that were developed during
the timeframe in which they were firstly introduced (QEMU 7.0) up to
this day, which means basically every other firmware that has RISC-V AIA
support.

The problem lies on the renamed properties, in particular the
'riscv,delegation' rename. The presence of the extra '#msi-cells' in the
IMSICS DT has also proven to be a problem.

This puts us in a bind because we need these changes to be compliant
with the dt-bindings spec - and new software should abide to them - and
at the same time we don't want to break existing firmwares and distros
in the wild.

There's no good solution here. The less unpleasant solution is to add a
compatibility flag for the 'virt' machine, 'legacy-aia-dt', that can be
used for older software to keep their AIA support working with the
'virt' machine when using QEMU 9.1 and newer. This will give them time
to adapt to the DT changes.

Note: the flag is default 'off', i.e. we'll use the correct AIA DT by
default.

Fixes: 29390fdbc1 ("hw/riscv/virt.c: rename aplic nodename to 'interrupt-controller'")
Fixes: b1f1e9dcfa ("hw/riscv/virt.c: aplic DT: rename prop to 'riscv, delegation'")
Fixes: e8ad5817b2 ("hw/riscv/virt.c: change imsic nodename to 'interrupt-controller'")
Fixes: f42cdf2ea5 ("hw/riscv/virt.c: imsics DT: add '#msi-cells'")
Signed-off-by: Daniel Henrique Barboza <dbarboza@ventanamicro.com>